Refactor actions panel into modular components
