+++ packages/web/src/components/actions/ActionsPanel.tsx
@@ -1,848 +1,19 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { getActionCosts, getActionRequirements } from '@kingdom-builder/engine';
-import {
-	RESOURCES,
-	POPULATION_ROLES,
-	POPULATION_INFO,
-	SLOT_INFO,
-	LAND_INFO,
-	type Focus,
-	type PopulationRoleId,
-} from '@kingdom-builder/contents';
+import React, { useEffect, useMemo, useState } from 'react';
+import { RESOURCES } from '@kingdom-builder/contents';
 import {
 	describeContent,
 	summarizeContent,
-	splitSummary,
 	type Summary,
 } from '../../translation';
-import ActionCard from './ActionCard';
 import { useGameEngine } from '../../state/GameContext';
 import { isActionPhaseActive } from '../../utils/isActionPhaseActive';
-import { getRequirementIcons } from '../../utils/getRequirementIcons';
 import { useAnimate } from '../../utils/useAutoAnimate';
-import GenericActions from './GenericActions';
-import { formatMissingResources } from './utils';
-import {
-	isHirablePopulation,
-	collectPopulationRolesFromEffects,
-	getPopulationIconFromRole,
-	type EffectConfig,
-	type PopulationRegistryLike,
-} from './populationHelpers';
-
-export interface Action {
-	id: string;
-	name: string;
-	system?: boolean;
-	order?: number;
-	category?: string;
-	focus?: Focus;
-	requirements?: unknown[];
-	effects?: unknown[];
-}
-interface Development {
-	id: string;
-	name: string;
-	system?: boolean;
-	order?: number;
-	focus?: Focus;
-}
-interface Building {
-	id: string;
-	name: string;
-	icon?: string;
-	focus?: Focus;
-}
-
-export type DisplayPlayer = ReturnType<
-	typeof useGameEngine
->['ctx']['activePlayer'];
-
-type RequirementConfig = {
-	type?: string;
-	method?: string;
-	params?: Record<string, unknown>;
-};
-
-type EvaluatorConfig = {
-	type?: string;
-	params?: Record<string, unknown>;
-};
-
-function getIconsFromEvaluator(
-	evaluator: EvaluatorConfig | undefined,
-	roleId: PopulationRoleId,
-	populations: PopulationRegistryLike,
-): string[] {
-	if (!evaluator || evaluator.type !== 'population') {
-		return [];
-	}
-	const params = evaluator.params ?? {};
-	const rawRole = params['role'];
-	if (typeof rawRole === 'string') {
-		if (rawRole.startsWith('$')) {
-			if (rawRole === '$role') {
-				const icon = getPopulationIconFromRole(roleId, populations);
-				return icon ? [icon] : [];
-			}
-			const placeholderIcon = getPopulationIconFromRole(
-				rawRole.slice(1),
-				populations,
-			);
-			return placeholderIcon ? [placeholderIcon] : [];
-		}
-		const icon = getPopulationIconFromRole(rawRole, populations);
-		return icon ? [icon] : [];
-	}
-	const genericIcon = POPULATION_INFO.icon;
-	return genericIcon ? [genericIcon] : [];
-}
-
-function determineRaisePopRoles(
-	actionDefinition: Action | undefined,
-	populations: PopulationRegistryLike,
-): PopulationRoleId[] {
-	const explicitRoles = new Set<string>();
-	const usesPlaceholder = collectPopulationRolesFromEffects(
-		(actionDefinition?.effects as EffectConfig[]) ?? [],
-		explicitRoles,
-	);
-	const orderedRoles = new Set<string>(explicitRoles);
-	if (usesPlaceholder || explicitRoles.size === 0) {
-		for (const [roleId, populationDef] of populations.entries()) {
-			if (!explicitRoles.has(roleId) && !isHirablePopulation(populationDef)) {
-				continue;
-			}
-			orderedRoles.add(roleId);
-		}
-	}
-	const result: PopulationRoleId[] = [];
-	for (const roleId of orderedRoles) {
-		try {
-			const population = populations.get(roleId);
-			if (!explicitRoles.has(roleId) && !isHirablePopulation(population)) {
-				continue;
-			}
-			result.push(roleId as PopulationRoleId);
-		} catch {
-			// Skip unknown population ids.
-		}
-	}
-	return result;
-}
-
-function playerHasRequiredResources(
-	player: DisplayPlayer,
-	costs: Record<string, number>,
-): boolean {
-	return Object.entries(costs).every(([resourceKey, requiredAmount]) => {
-		const playerResourceAmount = player.resources[resourceKey] || 0;
-		const neededAmount = requiredAmount ?? 0;
-		return playerResourceAmount >= neededAmount;
-	});
-}
-
-function sumNonActionCosts(
-	costs: Record<string, number>,
-	actionCostResource: string,
-): number {
-	return Object.entries(costs).reduce((sum, [resourceKey, requiredAmount]) => {
-		if (resourceKey === actionCostResource) {
-			return sum;
-		}
-		const neededAmount = requiredAmount ?? 0;
-		return sum + neededAmount;
-	}, 0);
-}
-
-function formatLandRequirement(prefix: string): string {
-	return [
-		`${prefix} ${LAND_INFO.icon} ${LAND_INFO.label}`,
-		`with free ${SLOT_INFO.icon} ${SLOT_INFO.label}`,
-	].join(' ');
-}
-
-function getOptionalProperty<T>(
-	value: unknown,
-	property: string,
-): T | undefined {
-	if (typeof value === 'object' && value !== null && property in value) {
-		return (value as Record<string, unknown>)[property] as T;
-	}
-	return undefined;
-}
-
-function buildRequirementIconsForRole(
-	actionDefinition: Action | undefined,
-	roleId: PopulationRoleId,
-	baseIcons: string[],
-	populations: PopulationRegistryLike,
-): string[] {
-	const icons = new Set(baseIcons);
-	const requirements = actionDefinition?.requirements as
-		| RequirementConfig[]
-		| undefined;
-	if (!requirements) {
-		return Array.from(icons).filter(Boolean);
-	}
-	for (const requirement of requirements) {
-		if (requirement.type !== 'evaluator' || requirement.method !== 'compare') {
-			continue;
-		}
-		const params = requirement.params ?? {};
-		const left = params['left'] as EvaluatorConfig | undefined;
-		const right = params['right'] as EvaluatorConfig | undefined;
-		for (const icon of getIconsFromEvaluator(left, roleId, populations)) {
-			icons.add(icon);
-		}
-		for (const icon of getIconsFromEvaluator(right, roleId, populations)) {
-			icons.add(icon);
-		}
-	}
-	return Array.from(icons).filter(Boolean);
-}
-
-function RaisePopOptions({
-	action,
-	player,
-	canInteract,
-}: {
-	action: Action;
-	player: DisplayPlayer;
-	canInteract: boolean;
-}) {
-	const {
-		ctx,
-		handlePerform,
-		handleHoverCard,
-		clearHoverCard,
-		actionCostResource,
-	} = useGameEngine();
-	const formatRequirement = (req: string) => req;
-	const populationRegistry = useMemo(
-		() => ctx.populations as unknown as PopulationRegistryLike,
-		[ctx.populations],
-	);
-	const actionDefinition = useMemo(
-		() => ctx.actions.get(action.id) as Action | undefined,
-		[ctx.actions, action.id],
-	);
-	const baseRequirementIcons = useMemo(
-		() => getRequirementIcons(action.id, ctx),
-		[action.id, ctx],
-	);
-	const roleOptions = useMemo(
-		() => determineRaisePopRoles(actionDefinition, populationRegistry),
-		[actionDefinition, populationRegistry],
-	);
-	const getRequirementIconsForRole = useCallback(
-		(role: PopulationRoleId) =>
-			buildRequirementIconsForRole(
-				actionDefinition,
-				role,
-				baseRequirementIcons,
-				populationRegistry,
-			),
-		[actionDefinition, baseRequirementIcons, populationRegistry],
-	);
-	return (
-		<>
-			{roleOptions.map((role) => {
-				const costsBag = getActionCosts(action.id, ctx);
-				const costEntries = Object.entries(costsBag);
-				const costs: Record<string, number> = {};
-				for (const [costKey, costAmount] of costEntries) {
-					costs[costKey] = costAmount ?? 0;
-				}
-				let upkeep: Record<string, number> | undefined;
-				try {
-					upkeep = populationRegistry.get(role)?.upkeep;
-				} catch {
-					upkeep = undefined;
-				}
-				const rawRequirements = getActionRequirements(action.id, ctx);
-				const requirements = rawRequirements.map(formatRequirement);
-				const canPay = playerHasRequiredResources(player, costs);
-				const meetsReq = requirements.length === 0;
-				const enabled = canPay && meetsReq && canInteract;
-				const requirementIcons = getRequirementIconsForRole(role);
-				const insufficientTooltip = formatMissingResources(
-					costs,
-					player.resources,
-				);
-				const title = !meetsReq
-					? requirements.join(', ')
-					: !canPay
-						? (insufficientTooltip ?? 'Cannot pay costs')
-						: undefined;
-				const summary = describeContent('action', action.id, ctx, {
-					role,
-				});
-				const shortSummary = summarizeContent('action', action.id, ctx, {
-					role,
-				});
-				return (
-					<ActionCard
-						key={role}
-						title={
-							<>
-								{ctx.actions.get(action.id).icon || ''}
-								{POPULATION_ROLES[role]?.icon} {ctx.actions.get(action.id).name}
-								: {POPULATION_ROLES[role]?.label}
-							</>
-						}
-						costs={costs}
-						upkeep={upkeep}
-						playerResources={player.resources}
-						actionCostResource={actionCostResource}
-						requirements={requirements}
-						requirementIcons={requirementIcons}
-						summary={shortSummary}
-						enabled={enabled}
-						tooltip={title}
-						focus={(ctx.actions.get(action.id) as Action | undefined)?.focus}
-						onClick={() => {
-							if (!canInteract) {
-								return;
-							}
-							void handlePerform(action, { role });
-						}}
-						onMouseEnter={() => {
-							const { effects, description } = splitSummary(summary);
-							handleHoverCard({
-								title: `${ctx.actions.get(action.id).icon || ''}${
-									POPULATION_ROLES[role]?.icon
-								} ${ctx.actions.get(action.id).name}: ${
-									POPULATION_ROLES[role]?.label || ''
-								}`,
-								effects,
-								requirements,
-								costs,
-								upkeep,
-								...(description && { description }),
-								bgClass:
-									'bg-gradient-to-br from-white/80 to-white/60 dark:from-slate-900/80 dark:to-slate-900/60',
-							});
-						}}
-						onMouseLeave={clearHoverCard}
-					/>
-				);
-			})}
-		</>
-	);
-}
-
-function BasicOptions({
-	actions,
-	summaries,
-	player,
-	canInteract,
-}: {
-	actions: Action[];
-	summaries: Map<string, Summary>;
-	player: DisplayPlayer;
-	canInteract: boolean;
-}) {
-	const listRef = useAnimate<HTMLDivElement>();
-	return (
-		<div>
-			<h3 className="font-medium">
-				Basic{' '}
-				<span className="italic text-sm font-normal">
-					(Effects take place immediately, unless stated otherwise)
-				</span>
-			</h3>
-			<div
-				ref={listRef}
-				className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-2 mt-1"
-			>
-				<GenericActions
-					actions={actions}
-					summaries={summaries}
-					player={player}
-					canInteract={canInteract}
-				/>
-			</div>
-		</div>
-	);
-}
-
-function HireOptions({
-	action,
-	player,
-	canInteract,
-}: {
-	action: Action;
-	player: DisplayPlayer;
-	canInteract: boolean;
-}) {
-	const listRef = useAnimate<HTMLDivElement>();
-	return (
-		<div>
-			<h3 className="font-medium">
-				Hire{' '}
-				<span className="italic text-sm font-normal">
-					(Recruit population instantly; upkeep and role effects apply while
-					they remain)
-				</span>
-			</h3>
-			<div
-				ref={listRef}
-				className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-2 mt-1"
-			>
-				<RaisePopOptions
-					action={action}
-					player={player}
-					canInteract={canInteract}
-				/>
-			</div>
-		</div>
-	);
-}
-
-function DevelopOptions({
-	action,
-	isActionPhase,
-	developments,
-	summaries,
-	hasDevelopLand,
-	player,
-	canInteract,
-}: {
-	action: Action;
-	isActionPhase: boolean;
-	developments: Development[];
-	summaries: Map<string, Summary>;
-	hasDevelopLand: boolean;
-	player: DisplayPlayer;
-	canInteract: boolean;
-}) {
-	const listRef = useAnimate<HTMLDivElement>();
-	const {
-		ctx,
-		handlePerform,
-		handleHoverCard,
-		clearHoverCard,
-		actionCostResource,
-	} = useGameEngine();
-	const landIdForCost = player.lands[0]?.id as string;
-	const entries = useMemo(() => {
-		return developments
-			.map((d) => {
-				const costsBag = getActionCosts(action.id, ctx, {
-					id: d.id,
-					landId: landIdForCost,
-				});
-				const costEntries = Object.entries(costsBag);
-				const costs: Record<string, number> = {};
-				for (const [costKey, costAmount] of costEntries) {
-					costs[costKey] = costAmount ?? 0;
-				}
-				const total = sumNonActionCosts(costs, actionCostResource);
-				return { d, costs, total };
-			})
-			.sort((firstEntry, secondEntry) => firstEntry.total - secondEntry.total);
-	}, [developments, ctx, action.id, landIdForCost, actionCostResource]);
-	return (
-		<div>
-			<h3 className="font-medium">
-				{ctx.actions.get(action.id)?.icon || ''}{' '}
-				{ctx.actions.get(action.id)?.name}{' '}
-				<span className="italic text-sm font-normal">
-					{'('}
-					Effects take place on build and last until development is removed
-					{')'}
-				</span>
-			</h3>
-			<div
-				ref={listRef}
-				className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-2 mt-1"
-			>
-				{entries.map(({ d, costs }) => {
-					const rawDevelopment = ctx.developments.get(d.id);
-					const upkeep = getOptionalProperty<Record<string, number>>(
-						rawDevelopment,
-						'upkeep',
-					);
-					const developLandRequirement = formatLandRequirement('Requires');
-					const requirements = hasDevelopLand ? [] : [developLandRequirement];
-					const canPay =
-						hasDevelopLand && playerHasRequiredResources(player, costs);
-					const summary = summaries.get(d.id);
-					const implemented = (summary?.length ?? 0) > 0;
-					// TODO: implement development effects
-					const enabled = canPay && isActionPhase && canInteract && implemented;
-					const developmentFocus = getOptionalProperty<Focus>(
-						rawDevelopment,
-						'focus',
-					);
-					const insufficientTooltip = formatMissingResources(
-						costs,
-						player.resources,
-					);
-					const missingLandTooltip = formatLandRequirement('No');
-					const title = !implemented
-						? 'Not implemented yet'
-						: !hasDevelopLand
-							? missingLandTooltip
-							: !canPay
-								? (insufficientTooltip ?? 'Cannot pay costs')
-								: undefined;
-					return (
-						<ActionCard
-							key={d.id}
-							title={
-								<>
-									{ctx.developments.get(d.id)?.icon} {d.name}
-								</>
-							}
-							costs={costs}
-							upkeep={upkeep}
-							playerResources={player.resources}
-							actionCostResource={actionCostResource}
-							requirements={requirements}
-							requirementIcons={[SLOT_INFO.icon]}
-							summary={summary}
-							implemented={implemented}
-							enabled={enabled}
-							tooltip={title}
-							focus={developmentFocus}
-							onClick={() => {
-								if (!canInteract) {
-									return;
-								}
-								const landId = player.lands.find(
-									(land) => land.slotsFree > 0,
-								)?.id;
-								void handlePerform(action, { id: d.id, landId });
-							}}
-							onMouseEnter={() => {
-								const full = describeContent('development', d.id, ctx);
-								const { effects, description } = splitSummary(full);
-								handleHoverCard({
-									title: `${ctx.actions.get(action.id)?.icon || ''} ${
-										ctx.actions.get(action.id)?.name
-									} - ${ctx.developments.get(d.id)?.icon || ''} ${d.name}`,
-									effects,
-									requirements,
-									costs,
-									upkeep,
-									...(description && { description }),
-									...(!implemented && {
-										description: 'Not implemented yet',
-										descriptionClass: 'italic text-red-600',
-									}),
-									bgClass:
-										'bg-gradient-to-br from-white/80 to-white/60 dark:from-slate-900/80 dark:to-slate-900/60',
-								});
-							}}
-							onMouseLeave={clearHoverCard}
-						/>
-					);
-				})}
-			</div>
-		</div>
-	);
-}
-
-function BuildOptions({
-	action,
-	isActionPhase,
-	buildings,
-	summaries,
-	descriptions,
-	player,
-	canInteract,
-}: {
-	action: Action;
-	isActionPhase: boolean;
-	buildings: Building[];
-	summaries: Map<string, Summary>;
-	descriptions: Map<string, Summary>;
-	player: DisplayPlayer;
-	canInteract: boolean;
-}) {
-	const listRef = useAnimate<HTMLDivElement>();
-	const {
-		ctx,
-		handlePerform,
-		handleHoverCard,
-		clearHoverCard,
-		actionCostResource,
-	} = useGameEngine();
-	const requirementIcons = useMemo(
-		() => getRequirementIcons(action.id, ctx),
-		[action.id, ctx],
-	);
-	const entries = useMemo(() => {
-		const owned = player.buildings;
-		return buildings
-			.filter((b) => !owned.has(b.id))
-			.map((b) => {
-				const costsBag = getActionCosts(action.id, ctx, {
-					id: b.id,
-				});
-				const costEntries = Object.entries(costsBag);
-				const costs: Record<string, number> = {};
-				for (const [costKey, costAmount] of costEntries) {
-					costs[costKey] = costAmount ?? 0;
-				}
-				const total = sumNonActionCosts(costs, actionCostResource);
-				return { b, costs, total };
-			})
-			.sort((firstEntry, secondEntry) => firstEntry.total - secondEntry.total);
-	}, [buildings, ctx, action.id, actionCostResource, player.buildings.size]);
-	return (
-		<div>
-			<h3 className="font-medium">
-				{ctx.actions.get(action.id)?.icon || ''}{' '}
-				{ctx.actions.get(action.id)?.name}{' '}
-				<span className="italic text-sm font-normal">
-					{'('}
-					Effects take place on build and last until building is removed
-					{')'}
-				</span>
-			</h3>
-			<div
-				ref={listRef}
-				className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-2 mt-1"
-			>
-				{entries.map(({ b, costs }) => {
-					const rawBuilding = ctx.buildings.get(b.id);
-					const buildingUpkeep = getOptionalProperty<Record<string, number>>(
-						rawBuilding,
-						'upkeep',
-					);
-					const buildingIcon = getOptionalProperty<string>(rawBuilding, 'icon');
-					const buildingFocus = getOptionalProperty<Focus>(
-						rawBuilding,
-						'focus',
-					);
-					const requirements = getActionRequirements(action.id, ctx);
-					const canPay = playerHasRequiredResources(player, costs);
-					const summary = summaries.get(b.id);
-					const implemented = (summary?.length ?? 0) > 0;
-					// TODO: implement building effects
-					const enabled = canPay && isActionPhase && canInteract && implemented;
-					const insufficientTooltip = formatMissingResources(
-						costs,
-						player.resources,
-					);
-					const title = !implemented
-						? 'Not implemented yet'
-						: !canPay
-							? (insufficientTooltip ?? 'Cannot pay costs')
-							: undefined;
-					const upkeep = buildingUpkeep;
-					return (
-						<ActionCard
-							key={b.id}
-							title={
-								<>
-									{buildingIcon || ''} {b.name}
-								</>
-							}
-							costs={costs}
-							upkeep={upkeep}
-							playerResources={player.resources}
-							actionCostResource={actionCostResource}
-							requirements={requirements}
-							requirementIcons={requirementIcons}
-							summary={summary}
-							implemented={implemented}
-							enabled={enabled}
-							tooltip={title}
-							focus={buildingFocus}
-							onClick={() => {
-								if (!canInteract) {
-									return;
-								}
-								void handlePerform(action, { id: b.id });
-							}}
-							onMouseEnter={() => {
-								const full = descriptions.get(b.id) ?? [];
-								const { effects, description } = splitSummary(full);
-								handleHoverCard({
-									title: `${ctx.actions.get(action.id)?.icon || ''} ${
-										ctx.actions.get(action.id)?.name
-									} - ${ctx.buildings.get(b.id)?.icon || ''} ${b.name}`,
-									effects,
-									requirements,
-									costs,
-									upkeep,
-									...(description && { description }),
-									...(!implemented && {
-										description: 'Not implemented yet',
-										descriptionClass: 'italic text-red-600',
-									}),
-									bgClass:
-										'bg-gradient-to-br from-white/80 to-white/60 dark:from-slate-900/80 dark:to-slate-900/60',
-								});
-							}}
-							onMouseLeave={clearHoverCard}
-						/>
-					);
-				})}
-			</div>
-		</div>
-	);
-}
-
-function DemolishOptions({
-	action,
-	isActionPhase,
-	player,
-	canInteract,
-}: {
-	action: Action;
-	isActionPhase: boolean;
-	player: DisplayPlayer;
-	canInteract: boolean;
-}) {
-	const listRef = useAnimate<HTMLDivElement>();
-	const {
-		ctx,
-		handlePerform,
-		handleHoverCard,
-		clearHoverCard,
-		actionCostResource,
-	} = useGameEngine();
-	const entries = useMemo(() => {
-		return Array.from(player.buildings)
-			.map((id) => {
-				const rawBuilding = ctx.buildings.get(id);
-				if (
-					typeof rawBuilding !== 'object' ||
-					rawBuilding === null ||
-					!('name' in rawBuilding)
-				) {
-					return null;
-				}
-				const building = rawBuilding as Building;
-				const costsBag = getActionCosts(action.id, ctx, {
-					id,
-				});
-				const costs: Record<string, number> = {};
-				const costEntries = Object.entries(costsBag);
-				for (const [costKey, costAmount] of costEntries) {
-					costs[costKey] = costAmount ?? 0;
-				}
-				const total = sumNonActionCosts(costs, actionCostResource);
-				return { id, building, costs, total };
-			})
-			.filter(
-				(
-					entry,
-				): entry is {
-					id: string;
-					building: Building;
-					costs: Record<string, number>;
-					total: number;
-				} => entry !== null,
-			)
-			.sort((firstEntry, secondEntry) => {
-				if (firstEntry.total !== secondEntry.total) {
-					return firstEntry.total - secondEntry.total;
-				}
-				return firstEntry.building.name.localeCompare(
-					secondEntry.building.name,
-				);
-			});
-	}, [ctx, action.id, actionCostResource, player.buildings.size]);
-
-	if (entries.length === 0) {
-		return null;
-	}
-
-	return (
-		<div>
-			<h3 className="font-medium">
-				{ctx.actions.get(action.id)?.icon || ''}{' '}
-				{ctx.actions.get(action.id)?.name}{' '}
-				<span className="italic text-sm font-normal">
-					{'('}
-					Removes a structure and its ongoing benefits
-					{')'}
-				</span>
-			</h3>
-			<div
-				ref={listRef}
-				className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-2 mt-1"
-			>
-				{entries.map(({ id, building, costs }) => {
-					const requirements: string[] = [];
-					const canPay = playerHasRequiredResources(player, costs);
-					const summary = summarizeContent('building', id, ctx, {
-						installed: true,
-					});
-					const implemented = (summary?.length ?? 0) > 0;
-					const enabled = canPay && isActionPhase && canInteract && implemented;
-					const buildingFocus = building.focus;
-					const insufficientTooltip = formatMissingResources(
-						costs,
-						player.resources,
-					);
-					const title = !implemented
-						? 'Not implemented yet'
-						: !canPay
-							? (insufficientTooltip ?? 'Cannot pay costs')
-							: undefined;
-					const upkeep = ctx.buildings.get(id)?.upkeep;
-					return (
-						<ActionCard
-							key={id}
-							title={
-								<>
-									{ctx.buildings.get(id)?.icon || ''} {building.name}
-								</>
-							}
-							costs={costs}
-							upkeep={upkeep}
-							playerResources={player.resources}
-							actionCostResource={actionCostResource}
-							requirements={requirements}
-							requirementIcons={[]}
-							summary={summary}
-							implemented={implemented}
-							enabled={enabled}
-							tooltip={title}
-							focus={buildingFocus}
-							onClick={() => {
-								if (!canInteract) {
-									return;
-								}
-								void handlePerform(action, { id });
-							}}
-							onMouseEnter={() => {
-								const full = describeContent('building', id, ctx, {
-									installed: true,
-								});
-								const { effects, description } = splitSummary(full);
-								handleHoverCard({
-									title: `${ctx.actions.get(action.id)?.icon || ''} ${
-										ctx.actions.get(action.id)?.name
-									} - ${ctx.buildings.get(id)?.icon || ''} ${building.name}`,
-									effects,
-									requirements,
-									costs,
-									upkeep,
-									...(description && { description }),
-									...(!implemented && {
-										description: 'Not implemented yet',
-										descriptionClass: 'italic text-red-600',
-									}),
-									bgClass:
-										'bg-gradient-to-br from-white/80 to-white/60 dark:from-slate-900/80 dark:to-slate-900/60',
-								});
-							}}
-							onMouseLeave={clearHoverCard}
-						/>
-					);
-				})}
-			</div>
-		</div>
-	);
-}
+import BasicOptions from './BasicOptions';
+import BuildOptions from './BuildOptions';
+import DemolishOptions from './DemolishOptions';
+import DevelopOptions from './DevelopOptions';
+import HireOptions from './HireOptions';
+import type { Action, Building, Development, DisplayPlayer } from './types';
 
 export default function ActionsPanel() {
 	const { ctx, tabsEnabled, actionCostResource } = useGameEngine();
@@ -868,7 +39,7 @@ export default function ActionsPanel() {
 		}
 	}, [isLocalTurn, viewingOpponent]);
 
-	const selectedPlayer = viewingOpponent ? opponent : player;
+	const selectedPlayer: DisplayPlayer = viewingOpponent ? opponent : player;
 	const canInteract = isLocalTurn && isActionPhase && !viewingOpponent;
 	const panelDisabled = !canInteract;
 
@@ -922,6 +93,7 @@ export default function ActionsPanel() {
 		);
 		return map;
 	}, [actions, ctx]);
+
 	const developmentSummaries = useMemo(() => {
 		const map = new Map<string, Summary>();
 		developmentOptions.forEach((developmentDefinition) =>
@@ -932,6 +104,7 @@ export default function ActionsPanel() {
 		);
 		return map;
 	}, [developmentOptions, ctx]);
+
 	const buildingSummaries = useMemo(() => {
 		const map = new Map<string, Summary>();
 		buildingOptions.forEach((buildingDefinition) =>
@@ -942,6 +115,7 @@ export default function ActionsPanel() {
 		);
 		return map;
 	}, [buildingOptions, ctx]);
+
 	const buildingDescriptions = useMemo(() => {
 		const map = new Map<string, Summary>();
 		buildingOptions.forEach((buildingDefinition) =>
@@ -953,13 +127,21 @@ export default function ActionsPanel() {
 		return map;
 	}, [buildingOptions, ctx]);
 
-	const hasDevelopLand = selectedPlayer.lands.some((l) => l.slotsFree > 0);
-	const developAction = actions.find((a) => a.category === 'development');
-	const buildAction = actions.find((a) => a.category === 'building');
-	const demolishAction = actions.find((a) => a.category === 'building_remove');
-	const raisePopAction = actions.find((a) => a.category === 'population');
+	const hasDevelopLand = selectedPlayer.lands.some(
+		(land) => land.slotsFree > 0,
+	);
+	const developAction = actions.find(
+		(action) => action.category === 'development',
+	);
+	const buildAction = actions.find((action) => action.category === 'building');
+	const demolishAction = actions.find(
+		(action) => action.category === 'building_remove',
+	);
+	const raisePopAction = actions.find(
+		(action) => action.category === 'population',
+	);
 	const otherActions = actions.filter(
-		(a) => (a.category ?? 'basic') === 'basic',
+		(action) => (action.category ?? 'basic') === 'basic',
 	);
 
 	const toggleLabel = viewingOpponent
@@ -999,7 +181,7 @@ export default function ActionsPanel() {
 						<button
 							type="button"
 							className="inline-flex h-10 w-10 items-center justify-center rounded-full border border-white/60 bg-white/70 text-lg font-semibold text-slate-700 shadow-md transition hover:bg-white/90 hover:text-slate-900 focus:outline-none focus-visible:ring-2 focus-visible:ring-amber-400 dark:border-white/10 dark:bg-slate-900/80 dark:text-slate-100 dark:hover:bg-slate-900"
-							onClick={() => setViewingOpponent((prev) => !prev)}
+							onClick={() => setViewingOpponent((previous) => !previous)}
 							aria-label={toggleLabel}
 						>
 							{viewingOpponent ? '←' : '→'}
