+++ packages/web/src/components/actions/populationHelpers.ts
@@ -1,7 +1,9 @@
 import {
+	POPULATION_INFO,
 	POPULATION_ROLES,
 	type PopulationRoleId,
 } from '@kingdom-builder/contents';
+import type { Action } from './types';
 
 export interface PopulationDefinition {
 	id: string;
@@ -28,6 +30,17 @@ export interface PopulationRegistryLike {
 	get(id: string): PopulationDefinition;
 	entries(): [string, PopulationDefinition][];
 }
+
+type RequirementConfig = {
+	type?: string;
+	method?: string;
+	params?: Record<string, unknown>;
+};
+
+type EvaluatorConfig = {
+	type?: string;
+	params?: Record<string, unknown>;
+};
 
 export function isHirablePopulation(
 	population: PopulationDefinition | undefined,
@@ -98,4 +111,96 @@ export function getPopulationIconFromRole(
 		// Ignore missing population entries when deriving icons.
 	}
 	return '';
+}
+
+function getIconsFromEvaluator(
+	evaluator: EvaluatorConfig | undefined,
+	roleId: PopulationRoleId,
+	populations: PopulationRegistryLike,
+): string[] {
+	if (!evaluator || evaluator.type !== 'population') {
+		return [];
+	}
+	const params = evaluator.params ?? {};
+	const rawRole = params['role'];
+	if (typeof rawRole === 'string') {
+		if (rawRole.startsWith('$')) {
+			if (rawRole === '$role') {
+				const icon = getPopulationIconFromRole(roleId, populations);
+				return icon ? [icon] : [];
+			}
+			const placeholderIcon = getPopulationIconFromRole(
+				rawRole.slice(1),
+				populations,
+			);
+			return placeholderIcon ? [placeholderIcon] : [];
+		}
+		const icon = getPopulationIconFromRole(rawRole, populations);
+		return icon ? [icon] : [];
+	}
+	const genericIcon = POPULATION_INFO.icon;
+	return genericIcon ? [genericIcon] : [];
+}
+
+export function determineRaisePopRoles(
+	actionDefinition: Action | undefined,
+	populations: PopulationRegistryLike,
+): PopulationRoleId[] {
+	const explicitRoles = new Set<string>();
+	const usesPlaceholder = collectPopulationRolesFromEffects(
+		(actionDefinition?.effects as EffectConfig[]) ?? [],
+		explicitRoles,
+	);
+	const orderedRoles = new Set<string>(explicitRoles);
+	if (usesPlaceholder || explicitRoles.size === 0) {
+		for (const [roleId, populationDef] of populations.entries()) {
+			if (!explicitRoles.has(roleId) && !isHirablePopulation(populationDef)) {
+				continue;
+			}
+			orderedRoles.add(roleId);
+		}
+	}
+	const result: PopulationRoleId[] = [];
+	for (const roleId of orderedRoles) {
+		try {
+			const population = populations.get(roleId);
+			if (!explicitRoles.has(roleId) && !isHirablePopulation(population)) {
+				continue;
+			}
+			result.push(roleId as PopulationRoleId);
+		} catch {
+			// Ignore missing population ids when collecting options.
+		}
+	}
+	return result;
+}
+
+export function buildRequirementIconsForRole(
+	actionDefinition: Action | undefined,
+	roleId: PopulationRoleId,
+	baseIcons: string[],
+	populations: PopulationRegistryLike,
+): string[] {
+	const icons = new Set(baseIcons);
+	const requirements = actionDefinition?.requirements as
+		| RequirementConfig[]
+		| undefined;
+	if (!requirements) {
+		return Array.from(icons).filter(Boolean);
+	}
+	for (const requirement of requirements) {
+		if (requirement.type !== 'evaluator' || requirement.method !== 'compare') {
+			continue;
+		}
+		const params = requirement.params ?? {};
+		const left = params['left'] as EvaluatorConfig | undefined;
+		const right = params['right'] as EvaluatorConfig | undefined;
+		for (const icon of getIconsFromEvaluator(left, roleId, populations)) {
+			icons.add(icon);
+		}
+		for (const icon of getIconsFromEvaluator(right, roleId, populations)) {
+			icons.add(icon);
+		}
+	}
+	return Array.from(icons).filter(Boolean);
 }
